Rework Limits and submodules
